Initial commit
This is the 1st commit of the YAML parser, yamil. In its current state
it is non-functional. I'm brushing up on my Ragel knowledge, it seems
like I've forgotten how to use this wonderful tools.

Also, it seems like YAML-files are quite more advanced than I have
realized before. I just thought that it was just a tad bit more complex
than JSON-files. So I will not aim at a fully conformant parser to
begin with. I will make it mostly just to fit my needs, but my intention
os of course to make it useful to others too.